+++ docs/changelog.md
@@ -1,4 +1,5 @@
 # TBD
+* Implement skeleton for the Startosis engine
 
 #### Fixes
 * Fixed a message that referred to an old repo.